api-versions/requirements.txt
azure-mgmt-resource==2.0.0